+++ tests/dummy/app/controllers/components-examples/flexberry-objectlistview/return-with-query-params/ember-flexberry-dummy-suggestion-return-with-query-params-list.js
@@ -1,0 +1,23 @@
+import ListFormController from 'ember-flexberry/controllers/list-form';
+import ListFormControllerOperationsIndicationMixin from 'dummy/mixins/list-form-controller-operations-indication';
+
+export default ListFormController.extend(ListFormControllerOperationsIndicationMixin, {
+  /**
+    Name of related edit form route.
+
+    @property editFormRoute
+    @type String
+    @default 'ember-flexberry-dummy-suggestion-edit'
+   */
+  editFormRoute: 'components-examples/flexberry-objectlistview/return-with-query-params/ember-flexberry-dummy-suggestion-return-with-query-params-edit',
+
+  exportExcelProjection: 'SuggestionL',
+
+  /**
+    Cout of list loading.
+
+    @property loadCount
+    @type Int
+  */
+  loadCount: 0,
+});