+++ tests/dummy/app/locales/en/translations.js
@@ -268,9 +268,9 @@ Ember.$.extend(true, translations, {
             'title': ''
           }
         },
-        lock: {
-          caption: 'Blocking',
-          title: 'Block list',
+        'lock': {
+          'caption': 'Blocking',
+          'title': 'Block list',
         },
         'components-examples': {
           'caption': 'Components examples',
@@ -470,6 +470,10 @@ Ember.$.extend(true, translations, {
               'caption': 'Columns markup',
               'title': ''
             },
+            'return-from-ediform': {
+              'title': 'Return from edit-form to list-form with queryParameter',
+              'return-button': 'Return'
+            },
             'lock-services-editor-view-list': {
               'caption': 'Example displaying username which the object was locked',
               'title': ''
@@ -504,6 +508,10 @@ Ember.$.extend(true, translations, {
             'title': '',
             'settings-example': {
               'caption': 'Settings example',
+              'title': ''
+            },
+            'settings-example-inner': {
+              'caption': 'Settings example (toggler in a toggler)',
               'title': ''
             },
             'ge-into-toggler-example': {
@@ -936,6 +944,11 @@ Ember.$.extend(true, translations, {
         'settings-example': {
           'caption': 'Flexberry-toggler. Settings example',
           'togglerContent': 'Some expandable/collapsable content'
+        },
+        'settings-example-inner': {
+          'caption': 'Flexberry-toggler. Settings example',
+          'togglerContent': 'Some expandable/collapsable content',
+          'innerTogglerContent': 'Some expandable/collapsable content in an inner toggler'
         },
         'ge-into-toggler-example': {
           'caption': 'Flexberry-toggler. GroupEdit into toggler example'